Fix 64-bit types for Wasm target, and make bitmap display nicer

+++ src/kernel/types.h
@@ -8,12 +8,12 @@
 typedef char i8;
 typedef short i16;
 typedef int i32;
-typedef long int i64;
+typedef long long int i64;
 
 typedef unsigned char u8;
 typedef unsigned short u16;
 typedef unsigned int u32;
-typedef unsigned long int u64;
+typedef unsigned long long int u64;
 
 typedef float f32;
 typedef double f64;
